server/requirements.txt
Flask
pymongo
python-dotenv
os
gridfs
io
